Moved browser tests out of the project root.
